moved types.ts folder into records folder from redux folder.
